backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.10.0
    # via
    #   django
    #   django-cors-headers
django==5.2.8
    # via
    #   -r requirements.in
    #   django-cors-headers
django-cors-headers==4.9.0
    # via -r requirements.in
iniconfig==2.3.0
    # via pytest
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via pytest
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via -r requirements.in
pytest==9.0.1
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.11.1
    # via -r requirements.in
sqlparse==0.5.3
    # via django
